diode-ping-agent/pyproject.toml
[project]
name = "diode-ping-agent"
version = "0.0.1"  # Overwritten during the build process
description = "NetBox Labs, Diode Ping Agent"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    {name = "NetBox Labs", email = "support@netboxlabs.com" }
]
maintainers = [
    {name = "NetBox Labs", email = "support@netboxlabs.com" }
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
]

dependencies = [
    "ping3==4.0.8",
    "netboxlabs-diode-sdk==0.2.0",
    "pydantic==2.7.1",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = ["black", "check-manifest", "ruff"]
test = ["coverage", "pytest", "pytest-cov"]

[project.urls]
"Homepage" = "https://netboxlabs.com/"

[project.scripts]
diode-napalm-agent = "diode_napalm.cli.cli:main"

[tool.setuptools]
packages = [
    "diode_ping",
]

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.ruff]
line-length = 140

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["C", "D", "E", "F", "I", "R", "UP", "W"]
ignore = ["F401", "D203", "D212", "D400", "D401", "D404", "RET504"]